I changes the projects structure by moving some logic from accounts applcation to offers app
